prepare for default input
