+++ src/pages/Explore/components/Filters/components/PriceFilter/styles.module.scss
@@ -1,0 +1,28 @@
+@import 'src/styles/helpers.scss';
+
+.prices {
+  @include flexbox(center, flex-start);
+  margin-left: 12px;
+}
+
+.box {
+  border: 1px solid $color-dark-3;
+  border-radius: 10px;
+  width: 160px;
+  height: 40px;
+  @include flexbox(center, flex-start);
+  & + & {
+    margin-left: 16px;
+  }
+}
+
+@include a {
+  .prices {
+    margin-left: 0;
+    margin-top: 8px;
+    width: 100%;
+    .box {
+      width: 100%;
+    }
+  }
+}
