+++ src/pages/Explore/components/Filters/components/index.ts
@@ -1,0 +1,2 @@
+
+export { default as PriceFilter } from './PriceFilter';
