+++ src/pages/Explore/components/Filters/styles.module.scss
@@ -37,7 +37,7 @@
 }
 
 .apply {
-  margin-left: 32px;
+  // margin-left: 32px;
 }
 
 @include d {
